[Refactor] Contents 페이지 컴포넌트 분리 (2023/09/18)
